requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
PyMySQL==1.1.0
Werkzeug==3.0.1
google-generativeai==0.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
python-dateutil==2.8.2